Refactor categoryFilter
